adding initial activity testing
To test an activity in isolation you can use the ActivityUnitTestCase
class. This class allows you to check the layout of the activity and to
check if intents are triggered as planned. The intents is not send to
the Android system but you can use the getStartedActivityIntent() method
to access a potential intent and validate its data.

ActivityUnitTestCase starts the activity in an IsolatedContext, i.e.
mainly isolated from the Android system.

ActivityUnitTestCase can be used to test layouts and isolated methods in
the activity.